% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beta.R
\name{create.beta}
\alias{create.beta}
\title{Creates matrix of beta coefficients.}
\usage{
create.beta(dat, ...)
}
\arguments{
\item{dat}{A data frame with columns \code{id}, \code{time} and \code{y1} to \code{y6} for the six channels.}
}
\value{
a list
}
\description{
Creates matrix of beta coefficients.
}
\examples{
# to do
}
